CATER FOR ZERO SIZE WEB APP DATABASE CREATED BY THE ADMIN SERVICE
The admin service will create an application database file of zero size
if one does not exist. The app service tests if the application database
file exists and, if it doesn't, restores it from the backup. So, the app
service also must restore the database from backup if it exists already
but is of zero size.