BRING THE IMAGE HANDLING PATHS BACK TO THE ENVIRONMENT CONFIGURATION
I had moved these in to the application configuration on the basis that
they don't vary across environments. However, this caused an issue
because I had forgotten that it is only the environment configuration
that is made available to templates and these variables are used in
templates. So, they've been moved back here.